:hammer: remove secret checksums from Harbor deployments
The Harbor chart generates a bunch of secrets in Helm, which results in
the secret checksums changing every run (even though the real secrets
won't change since those are stored separately in Vault). Instead of
trying to get the chart to not, remove all checksums related to secrets
(it seems that the keys in question on the secrets are not even used in
our configuration).