:screwdriver: allow ArgoCD HAProxy deployment to progress
Previously, due to the node antiaffinity on HAProxy pods combined with
our schedulable node count exactly equaling the replica count, ArgoCD's
(internal) HAProxy deployment would not be able to progress since there
are no nodes to schedule the new pods on. Adjust the rolling update
strategy to limit the max pods to 3 at a time.