:sparkles: deploy OCF Object Ocean (O3)
OCF Object Ocean is an HDD-backed S3-compatible object storage system
with metadata stored on NVMe drives for better latency on meta
operations. The plan is to use this for all object storage needs, and to
take advantage of multitenancy to eventually offer this as a service to
the broader campus community.