:wastebasket: remove istio and knative
Both of these are massively complex things that want to replace
everything else that we've built. I've spent the better part of three
days configuring them, and while they sort of work, there are just so
many remaining issues to debug.

Instead, what I think the OCF actually needs is a lightweight way to
deploy a piece of code and have it interact with RabbitMQ. We don't
really need the staged rollouts and chaos engineering and things that
Knative provides, even though they are cool.

That said, sometimes the only reason we do a thing at the OCF is because
it is cool and we want to learn-- so maybe I'll revisit it at some
point.