:bug: fix remote backups
We forgot to enable MoveData to actually copy the snapshot to the remote. Right now this just keeps a snapshot locally, which is pretty good, but we also want remote backups!

Not merging by myself because I'm not sure if this deletes the local VolumeSnapshots... Maybe it's okay if it does?